a demonstration of an appropriate use of a binary tree in C++
The purpose of this program is to demonstrate an appropriate use of a
binary tree in C++. I was provided with a file containing a list of
employee names and their ID numbers. This program fills a binary tree
with the employee information and allows the user to search for the name
of an employee by inputting that employee's ID number.